+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Simplify foreach with `in_array` function (#830)
 * Typehinted return value on Expectation#verify. (#832)
 * Fix shouldNotHaveReceived with HigherOrderMessage (#842)
+* Deprecates shouldDeferMissing (#839)
 
 ## 1.0.0 (2017-09-06)